requirements.txt
pyserial==3.5
ttkbootstrap~=1.10.1